add github repo card
